Removed test comments
OK, I made a change to the repository (removing test comments) without
creating a new branch. GitHub for Windows noticed the change and is
suggesting that I commit the change. So it appears that changes can be
committed directly to the master branch, and I assume, pushed to GitHub
in the cloud.